Move server types into server module
